requirements.txt
3==0.0.0
aniso8601==7.0.0
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.4
dotenv-python==0.0.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
